feat: Decimal money migration (Plan 13 Phase 2)
Replace binary `float` with fixed-precision `Decimal` / SQL `NUMERIC` for
every authoritative monetary value, so repeated aggregation is
order-independent and no binary-float drift reaches budget comparisons.

- domain/money.py: one seam for money precision — MONEY_SCALE=12,
  ROUND_HALF_EVEN, NUMERIC(24,12), and money() building Decimals from
  strings/ints (never Decimal(float)).
- Decimalize RateCard/BillableUsage/compute_cost, Model rates + image_prices,
  UsageEvent/Budget/PendingBudgetAlert/aggregates, RoutingDecisionRecord cost
  fields, the UsageMeter reservation/settlement/InFlightSpend path, spend_since,
  and the routing/cache savings deltas.
- ORM money columns -> NUMERIC(24,12); Alembic migration e5f6a7b8c9d0 converts
  float->numeric with `USING col::numeric(24,12)` (Postgres rounds each stored
  double to the intended decimal at scale 12), rehearsed via just test-postgres.
- API keeps JSON numbers: DTOs stay float and widen Decimal->float at the
  response boundary (compatibility serializer), so the OpenAPI schema and the
  number-typed console are unchanged. image_prices persist as decimal strings
  in JSON and re-parse to Decimal.
- Tests tightened to exact Decimal assertions; adds an order-independence proof
  (byte-identical totals across orders) and a Postgres float->numeric rehearsal.

+++ src/litestar_gateway/application/playground_service.py
@@ -233,4 +233,6 @@ def _cost(model: Model, prompt: object, completion: object) -> float | None:
         return None
     if not isinstance(prompt, int) or not isinstance(completion, int):
         return None
-    return prompt * model.input_cost_per_token + completion * model.output_cost_per_token
+    # Rates are exact Decimals (Plan 13 Phase 2); this is a display-only estimate
+    # (real billing runs through the meter), so widen to a JSON number here.
+    return float(prompt * model.input_cost_per_token + completion * model.output_cost_per_token)
