+++ tests/budgets/test_failover_orchestration.py
@@ -10,6 +10,7 @@ logical request consumes exactly one team-RPM hit across every attempt.
 from __future__ import annotations
 
 from datetime import UTC, datetime
+from decimal import Decimal
 from types import SimpleNamespace
 from typing import Any
 from uuid import UUID, uuid4
@@ -31,6 +32,7 @@ from litestar_gateway.domain.exceptions import (
     UpstreamResponseInvalid,
     UpstreamUnavailable,
 )
+from litestar_gateway.domain.money import money
 from litestar_gateway.domain.ports.rate_limiter import RateLimitDecision
 from litestar_gateway.domain.routing import (
     CandidateModel,
@@ -56,8 +58,8 @@ def _model(name: str) -> Model:
         params={},
         params_enforced={},
         api_version=None,
-        input_cost_per_token=0.01,
-        output_cost_per_token=0.01,
+        input_cost_per_token=money(0.01),
+        output_cost_per_token=money(0.01),
         enabled=True,
         created_at=datetime.now(UTC),
     )
@@ -67,7 +69,7 @@ def _budget(limit: float) -> Budget:
     return Budget(
         id=uuid4(),
         team_id=TEAM_ID,
-        limit_cost=limit,
+        limit_cost=money(limit),
         window=BudgetWindow.MONTHLY,
         created_at=datetime.now(UTC),
     )
@@ -122,7 +124,7 @@ class FakeCredentials:
 class FakeUsage:
     def __init__(self, spent: float = 0.0) -> None:
         self.events: list[Any] = []
-        self.spent = spent
+        self.spent = money(spent)
 
     async def record(self, event: Any) -> None:
         self.events.append(event)
@@ -130,7 +132,7 @@ class FakeUsage:
     async def enqueue_pending(self, event: Any) -> None:  # pragma: no cover
         raise AssertionError("outbox must not be used in these tests")
 
-    async def spend_since(self, team_id: UUID, since: datetime) -> float:
+    async def spend_since(self, team_id: UUID, since: datetime) -> Decimal:
         return self.spent
